* Game::transformItem works now 100% properly (if doesn't, please report ASAP, but I am running out of my powers with it) * Fixed a bug not updating shop window while moving item from inventory to backpack * Fix to npcsystem making able to sell stackable items properly (thanks Cykotitan for pointing out, I have to patch yet 0.2 npcsystem changes but have to take a look on it first) * Fixed compilation bug
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@4349 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ iologindata.cpp
@@ -1079,7 +1079,7 @@ bool IOLoginData::savePlayer(Player* player, bool preSave/* = true*/, bool shall
 			continue;
 
 		if(!g_config.getBool(ConfigManager::VIPLIST_PER_PLAYER))
-			sprintf(buffer, "%d, %d, %d", player->getAccount(), g_config.getNumber(ConfigManager::WORLD_ID), *it);
+			sprintf(buffer, "%d, %d, %d", player->getAccount(), (int32_t)g_config.getNumber(ConfigManager::WORLD_ID), *it);
 		else
 			sprintf(buffer, "%d, %d", player->getGUID(), *it);
